# api/swagger.yaml
openapi: 3.0.0
info:
  title: Poseidon API
  description: |
    This API is used by CodeOcean to run code in runners.
  version: '0.2.2'
components:
...
  responses:
    BadRequest:
      description: Request is invalid. E.g. request body does not follow the json schema required by the given route or url parameters are invalid.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                description: Explanation on why the request is invalid
                type: string
    Unauthorized:
      description: Client could not be authenticated
    NotFound:
      description: The entity with the given identifier does not exist.
    InternalServerError:
      description: Request could not be handled
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                description: Explanation on why the request could not be handled
                type: string
                example: Nomad server unreachable
              errorCode:
                description: Machine readable error description
                type: string
                enum:
                  - NOMAD_UNREACHABLE
                  - NOMAD_OVERLOAD
                  - NOMAD_INTERNAL_SERVER_ERROR
                  - UNKNOWN
                example: NOMAD_UNREACHABLE
...
paths:
...
  /runners/{runnerId}/files:
    patch:
      summary: Manipulate runner file system
      description: Delete the files with the given paths from the file system of the specified runner. Afterwards, copy the enclosed files to the runner. Existing files get overwritten and results of previous file copy operations on the same runner are present when executing multiple requests.
      tags:
        - runner
      parameters:
        - name: runnerId
          in: path
          schema:
            description: Runner on which the files should be placed
            type: string
            example: 123e4567-e89b-12d3-a456-426614174000
          required: true
      requestBody:
        description: Files to copy or delete
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                delete:
                  description: Array of filepaths that should be deleted. Each of the given files or directories should be deleted recursively.
                  type: array
                  items:
                    description: Location of the file or directory that should be deleted. Can be absolute (starting with /) or relative to the workspace directory.
                    type: string
                    example: /workspace
                copy:
                  description: Array of files that should be placed in the runner.
                  type: array
                  items:
                    type: object
                    properties:
                      path:
                        description: Location where the file should be placed. Can be absolute (starting with /) or relative to the workspace directory. Missing parent directories are created. If this ends with a /, the path is interpreted as a directory and content is ignored
                        type: string
                        example: /etc/passwd
                      content:
                        description: The content of the file. MUST be base64 encoded. If this is not given, the file is created with no content.
                        type: string
                        example: cm9vdDp4OjA6MDo6L3Jvb3Q6L2Jpbi9iYXNo # root:x:0:0::/root:/bin/bash
                    required:
                      - path
                    additionalProperties: false
              additionalProperties: false
      responses:
        "204":
          description: All files were saved
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /runners/{runnerId}/execute:
    post:
      summary: Execute a command
      description: Execute a command in the runner. Whether this starts the actual execution or only prepares a Websocket URL to start it depends on the implementation.
      tags:
        - runner
      parameters:
        - name: runnerId
          description: Runner on which the command should be executed
          in: path
          schema:
            type: string
            example: 123e4567-e89b-12d3-a456-426614174000
          required: true
      requestBody:
        description: Description what and how to execute
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                command:
                  description: The command to be executed. The working directory for this execution is the working directory of the image of the execution environment
                  type: string
                  example: python exercise.py
                environment:
                  description: Environment variables for this execution. The keys of this object are the variable names and the value of each key is the value of the variable with the same name
                  type: object
                  additionalProperties:
                    type: string
                    pattern: "[a-zA-Z_][a-zA-Z0-9_]+"
                  default: {}
                  example:
                    PATH: /bin
                timeLimit:
                  description: Specifies the time in seconds until this execution should be killed. 0 means do not kill
                  type: integer
                  default: 0
                  example: 5
              required:
                - command
              additionalProperties: false
      responses:
        "200":
          description: Success. Returns a Websocket URL to connect to
          content:
            application/json:
              schema:
                type: object
                properties:
                  websocketUrl:
                    description: A Websocket endpoint to connect to communicate with the process running in the runner
                    type: string
                    example: "ws://ws.example.com/path/to/websocket"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...